pushing current state
