+++ src/main/java/horse/wtf/nzyme/NzymeTrackerImpl.java
@@ -44,7 +44,7 @@ import horse.wtf.nzyme.dot11.probes.Dot11MonitorProbe;
 import horse.wtf.nzyme.dot11.probes.Dot11Probe;
 import horse.wtf.nzyme.dot11.probes.Dot11ProbeConfiguration;
 import horse.wtf.nzyme.notifications.Notification;
-import horse.wtf.nzyme.statistics.Statistics;
+import horse.wtf.nzyme.processing.FrameProcessor;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -67,7 +67,8 @@ public class NzymeTrackerImpl implements NzymeTracker {
     private final GroundStation groundStation;
     private final TrackerBanditManager banditManager;
     private final TrackerStateWatchdog trackerStateWatchdog;
-    private final Statistics statistics;
+
+    private final FrameProcessor frameProcessor;
 
     private final List<Dot11Probe> probes;
     private final MetricRegistry metrics;
@@ -85,10 +86,11 @@ public class NzymeTrackerImpl implements NzymeTracker {
 
         this.anonymizer = new Anonymizer(baseConfiguration.anonymize(), baseConfiguration.dataDirectory());
 
+        this.frameProcessor = new FrameProcessor();
+
         this.probes = Lists.newArrayList();
         this.hids = Lists.newArrayList();
 
-        this.statistics = new Statistics(this);
         this.metrics = new MetricRegistry();
 
         this.om = new ObjectMapper()
@@ -212,7 +214,7 @@ public class NzymeTrackerImpl implements NzymeTracker {
                     m.channelHopCommand(),
                     null,
                     null
-            ), metrics, statistics, anonymizer, true);
+            ), frameProcessor, metrics, anonymizer, this,true);
 
             probe.onChannelSwitch((previousChannel, newChannel) -> {
                 for (TrackerHID hid : hids) {
@@ -221,12 +223,11 @@ public class NzymeTrackerImpl implements NzymeTracker {
 
             });
 
-            // Register the bandit interceptor.
-            probe.addFrameInterceptors(new BanditIdentifierInterceptorSet(getBanditManager()).getInterceptors());
-
             probeExecutor.submit(probe.loop());
             this.probes.add(probe);
         }
+
+        frameProcessor.registerDot11Interceptors(new BanditIdentifierInterceptorSet(getBanditManager()).getInterceptors());
     }
 
     @Override
