+++ src/main/java/horse/wtf/nzyme/dot11/frames/Dot11BeaconFrame.java
@@ -19,6 +19,7 @@ package horse.wtf.nzyme.dot11.frames;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.google.auto.value.AutoValue;
+import horse.wtf.nzyme.dot11.Dot11FrameSubtype;
 import horse.wtf.nzyme.dot11.Dot11MetaInformation;
 import horse.wtf.nzyme.dot11.Dot11TaggedParameters;
 
@@ -29,18 +30,13 @@ public abstract class Dot11BeaconFrame implements Dot11Frame {
 
     @Nullable
     public abstract String ssid();
-
     public abstract String transmitter();
-
     public abstract String transmitterFingerprint();
-
     public abstract Dot11TaggedParameters taggedParameters();
-
     public abstract Dot11MetaInformation meta();
-
     public abstract byte[] payload();
-
     public abstract byte[] header();
+    public abstract byte frameType();
 
     @JsonIgnore
     public String descriptionString() {
@@ -65,6 +61,7 @@ public abstract class Dot11BeaconFrame implements Dot11Frame {
                 .meta(meta)
                 .payload(payload)
                 .header(header)
+                .frameType(Dot11FrameSubtype.BEACON)
                 .build();
     }
 
@@ -87,6 +84,8 @@ public abstract class Dot11BeaconFrame implements Dot11Frame {
         public abstract Builder payload(byte[] payload);
 
         public abstract Builder header(byte[] header);
+
+        public abstract Builder frameType(byte type);
 
         public abstract Dot11BeaconFrame build();
     }
