+++ src/main/java/horse/wtf/nzyme/dot11/frames/Dot11Frame.java
@@ -21,6 +21,7 @@ import horse.wtf.nzyme.dot11.Dot11MetaInformation;
 
 public interface Dot11Frame {
 
+    byte frameType();
     Dot11MetaInformation meta();
     byte[] payload();
     byte[] header();
