+++ src/main/java/horse/wtf/nzyme/dot11/frames/Dot11FrameFactory.java
@@ -1,0 +1,74 @@
+/*
+ * This file is part of nzyme.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the Server Side Public License, version 1,
+ * as published by MongoDB, Inc.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * Server Side Public License for more details.
+ *
+ * You should have received a copy of the Server Side Public License
+ * along with this program. If not, see
+ * <http://www.mongodb.com/licensing/server-side-public-license>.
+ */
+
+package horse.wtf.nzyme.dot11.frames;
+
+import com.codahale.metrics.MetricRegistry;
+import horse.wtf.nzyme.dot11.Dot11FrameSubtype;
+import horse.wtf.nzyme.dot11.Dot11MetaInformation;
+import horse.wtf.nzyme.dot11.MalformedFrameException;
+import horse.wtf.nzyme.dot11.anonymization.Anonymizer;
+import horse.wtf.nzyme.dot11.parsers.*;
+import org.pcap4j.packet.IllegalRawDataException;
+import org.pcap4j.packet.namednumber.Dot11FrameType;
+
+public class Dot11FrameFactory {
+
+    private final Dot11BeaconFrameParser beaconParser;
+    private final Dot11AssociationRequestFrameParser associationRequestParser;
+    private final Dot11AssociationResponseFrameParser associationResponseParser;
+    private final Dot11ProbeRequestFrameParser probeRequestParser;
+    private final Dot11ProbeResponseFrameParser probeResponseFrameParser;
+    private final Dot11DisassociationFrameParser disassociationParser;
+    private final Dot11AuthenticationFrameParser authenticationFrameParser;
+    private final Dot11DeauthenticationFrameParser deauthenticationFrameParser;
+
+    public Dot11FrameFactory(MetricRegistry metrics, Anonymizer anonymizer) {
+        beaconParser = new Dot11BeaconFrameParser(metrics, anonymizer);
+        associationRequestParser = new Dot11AssociationRequestFrameParser(metrics, anonymizer);
+        associationResponseParser = new Dot11AssociationResponseFrameParser(metrics, anonymizer);
+        probeRequestParser = new Dot11ProbeRequestFrameParser(metrics, anonymizer);
+        probeResponseFrameParser = new Dot11ProbeResponseFrameParser(metrics, anonymizer);
+        disassociationParser = new Dot11DisassociationFrameParser(metrics, anonymizer);
+        authenticationFrameParser = new Dot11AuthenticationFrameParser(metrics, anonymizer);
+        deauthenticationFrameParser = new Dot11DeauthenticationFrameParser(metrics, anonymizer);
+    }
+
+    public Dot11Frame build(Dot11FrameType type, byte[] payload, byte[] header, Dot11MetaInformation meta) throws MalformedFrameException, IllegalRawDataException {
+        switch(type.value()) {
+            case Dot11FrameSubtype.ASSOCIATION_REQUEST:
+                return associationRequestParser.parse(payload, header, meta);
+            case Dot11FrameSubtype.ASSOCIATION_RESPONSE:
+                return associationResponseParser.parse(payload, header, meta);
+            case Dot11FrameSubtype.PROBE_REQUEST:
+                return probeRequestParser.parse(payload, header, meta);
+            case Dot11FrameSubtype.PROBE_RESPONSE:
+                return probeResponseFrameParser.parse(payload, header, meta);
+            case Dot11FrameSubtype.BEACON:
+                return beaconParser.parse(payload, header, meta);
+            case Dot11FrameSubtype.DISASSOCIATION:
+                return disassociationParser.parse(payload, header, meta);
+            case Dot11FrameSubtype.AUTHENTICATION:
+                return authenticationFrameParser.parse(payload, header, meta);
+            case Dot11FrameSubtype.DEAUTHENTICATION:
+                return deauthenticationFrameParser.parse(payload, header, meta);
+            default:
+                throw new IllegalStateException("Unexpected value: " + type.value());
+        }
+    }
+
+}
