--- src/main/java/horse/wtf/nzyme/dot11/handlers/Dot11ProbeResponseFrameHandler.java
@@ -1,72 +1,0 @@
-/*
- * This file is part of nzyme.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the Server Side Public License, version 1,
- * as published by MongoDB, Inc.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * Server Side Public License for more details.
- *
- * You should have received a copy of the Server Side Public License
- * along with this program. If not, see
- * <http://www.mongodb.com/licensing/server-side-public-license>.
- */
-
-package horse.wtf.nzyme.dot11.handlers;
-
-import horse.wtf.nzyme.RemoteConnector;
-import horse.wtf.nzyme.dot11.frames.Dot11ProbeResponseFrame;
-import horse.wtf.nzyme.dot11.probes.Dot11Probe;
-import horse.wtf.nzyme.notifications.FieldNames;
-import horse.wtf.nzyme.notifications.Notification;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-
-public class Dot11ProbeResponseFrameHandler extends Dot11FrameHandler<Dot11ProbeResponseFrame> {
-
-    private static final Logger LOG = LogManager.getLogger(Dot11ProbeResponseFrameHandler.class);
-
-    private final RemoteConnector remote;
-
-    public Dot11ProbeResponseFrameHandler(Dot11Probe probe, RemoteConnector remote) {
-        super(probe);
-
-        this.remote = remote;
-    }
-
-    @Override
-    protected void doHandle(Dot11ProbeResponseFrame frame) {
-        String message;
-        if (frame.ssid() == null) {
-            message = frame.transmitter() + " responded to broadcast probe request from " + frame.destination();
-        } else {
-            message = frame.transmitter() + " responded to probe request from " + frame.destination() + " for " + frame.ssid();
-        }
-
-        remote.notifyUplinks(
-                new Notification(message, frame.meta().getChannel())
-                        .addField(FieldNames.DESTINATION, frame.destination())
-                        .addField(FieldNames.TRANSMITTER, frame.transmitter())
-                        .addField(FieldNames.SSID, frame.ssid() == null ? "[no SSID]" : frame.ssid())
-                        .addField(FieldNames.SECURITY_FULL, frame.taggedParameters().getFullSecurityString())
-                        .addField(FieldNames.IS_WPA1, frame.taggedParameters().isWPA1())
-                        .addField(FieldNames.IS_WPA2, frame.taggedParameters().isWPA2())
-                        .addField(FieldNames.IS_WPS, frame.taggedParameters().isWPS())
-                        .addField(FieldNames.SUBTYPE, "probe-resp"),
-                        frame.meta()
-        );
-
-        remote.forwardFrame(frame);
-
-        LOG.debug(message);
-    }
-
-    @Override
-    public String getName() {
-        return "probe-resp";
-    }
-
-}
