+++ src/main/java/horse/wtf/nzyme/dot11/interceptors/UnexpectedBSSIDInterceptorSet.java
@@ -50,7 +50,7 @@ public class UnexpectedBSSIDInterceptorSet {
         // React on probe-resp frames with unexpected BSSID.
         interceptors.add(new Dot11FrameInterceptor<Dot11ProbeResponseFrame>() {
             @Override
-            public void intercept(Dot11ProbeResponseFrame frame) throws IllegalRawDataException {
+            public void intercept(Dot11ProbeResponseFrame frame) {
                 // Don't consider broadcast frames.
                 if (frame.ssid() == null) {
                     return;
@@ -91,7 +91,7 @@ public class UnexpectedBSSIDInterceptorSet {
         // React on beacon frames with unexpected BSSID.
         interceptors.add(new Dot11FrameInterceptor<Dot11BeaconFrame>() {
             @Override
-            public void intercept(Dot11BeaconFrame frame) throws IllegalRawDataException {
+            public void intercept(Dot11BeaconFrame frame) {
                 // Don't consider broadcast frames.
                 if (frame.ssid() == null) {
                     return;
