+++ src/main/java/horse/wtf/nzyme/dot11/interceptors/UnexpectedSSIDInterceptorSet.java
@@ -50,7 +50,7 @@ public class UnexpectedSSIDInterceptorSet {
         // React on probe-resp frames from one of our BSSIDs that is advertising a network that is not ours.
         interceptors.add(new Dot11FrameInterceptor<Dot11ProbeResponseFrame>() {
             @Override
-            public void intercept(Dot11ProbeResponseFrame frame) throws IllegalRawDataException {
+            public void intercept(Dot11ProbeResponseFrame frame) {
                 // Don't consider broadcast frames.
                 if (frame.ssid() == null) {
                     return;
@@ -87,7 +87,7 @@ public class UnexpectedSSIDInterceptorSet {
         // React on beacon frames from one of our BSSIDs that is advertising a network that is not ours.
         interceptors.add(new Dot11FrameInterceptor<Dot11BeaconFrame>() {
             @Override
-            public void intercept(Dot11BeaconFrame frame) throws IllegalRawDataException {
+            public void intercept(Dot11BeaconFrame frame) {
                 // Don't consider broadcast frames.
                 if (frame.ssid() == null) {
                     return;
