+++ src/main/java/horse/wtf/nzyme/dot11/probes/Dot11MonitorProbe.java
@@ -21,34 +21,27 @@ import com.codahale.metrics.Meter;
 import com.codahale.metrics.MetricRegistry;
 import com.codahale.metrics.Timer;
 import com.google.common.base.Joiner;
-import com.google.common.collect.Lists;
 import horse.wtf.nzyme.RemoteConnector;
-import horse.wtf.nzyme.alerts.Alert;
 import horse.wtf.nzyme.bandits.trackers.trackerlogic.ChannelDesignator;
 import horse.wtf.nzyme.channels.ChannelHopper;
-import horse.wtf.nzyme.dot11.Dot11FrameInterceptor;
-import horse.wtf.nzyme.dot11.Dot11FrameSubtype;
 import horse.wtf.nzyme.dot11.Dot11MetaInformation;
 import horse.wtf.nzyme.dot11.MalformedFrameException;
 import horse.wtf.nzyme.dot11.anonymization.Anonymizer;
-import horse.wtf.nzyme.dot11.frames.*;
-import horse.wtf.nzyme.dot11.handlers.*;
-import horse.wtf.nzyme.dot11.parsers.*;
-import horse.wtf.nzyme.statistics.Statistics;
+import horse.wtf.nzyme.dot11.frames.Dot11FrameFactory;
+import horse.wtf.nzyme.notifications.FieldNames;
+import horse.wtf.nzyme.notifications.Notification;
+import horse.wtf.nzyme.processing.FrameProcessor;
 import horse.wtf.nzyme.util.MetricNames;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.joda.time.DateTime;
 import org.pcap4j.core.*;
-import org.pcap4j.packet.IllegalRawDataException;
 import org.pcap4j.packet.Packet;
 import org.pcap4j.packet.RadiotapPacket;
 import org.pcap4j.packet.namednumber.Dot11FrameType;
 
 import javax.annotation.Nullable;
 import java.io.EOFException;
-import java.util.Collections;
-import java.util.List;
 import java.util.concurrent.TimeoutException;
 import java.util.concurrent.atomic.AtomicBoolean;
 
@@ -61,21 +54,10 @@ public class Dot11MonitorProbe extends Dot11Probe {
     private PcapHandle pcap;
     private final ChannelHopper channelHopper;
     private final ChannelDesignator channelDesignator;
+    private final RemoteConnector remote;
 
-    private final Statistics statistics;
-
-    // Interceptors.
-    private final List<Dot11FrameInterceptor> frameInterceptors;
-
-    // Parsers.
-    private final Dot11BeaconFrameParser beaconParser;
-    private final Dot11AssociationRequestFrameParser associationRequestParser;
-    private final Dot11AssociationResponseFrameParser associationResponseParser;
-    private final Dot11ProbeRequestFrameParser probeRequestParser;
-    private final Dot11ProbeResponseFrameParser probeResponseFrameParser;
-    private final Dot11DisassociationFrameParser disassociationParser;
-    private final Dot11AuthenticationFrameParser authenticationFrameParser;
-    private final Dot11DeauthenticationFrameParser deauthenticationFrameParser;
+    private final Dot11FrameFactory frameFactory;
+    private final FrameProcessor frameProcessor;
 
     // Metrics
     private final Meter globalFrameMeter;
@@ -86,23 +68,14 @@ public class Dot11MonitorProbe extends Dot11Probe {
 
     private DateTime mostRecentFrameTimestamp;
 
-    public Dot11MonitorProbe(Dot11ProbeConfiguration configuration, MetricRegistry metrics, Statistics statistics, Anonymizer anonymizer, boolean hasDesignator) {
-        super(configuration, statistics, metrics);
+    public Dot11MonitorProbe(Dot11ProbeConfiguration configuration, FrameProcessor frameProcessor, MetricRegistry metrics, Anonymizer anonymizer, RemoteConnector remote, boolean hasDesignator) {
+        super(configuration, metrics);
 
-        this.statistics = statistics;
-
+        this.remote = remote;
         this.configuration = configuration;
-        this.frameInterceptors = Lists.newArrayList();
 
-        // Parsers.
-        beaconParser = new Dot11BeaconFrameParser(metrics, anonymizer);
-        associationRequestParser = new Dot11AssociationRequestFrameParser(metrics, anonymizer);
-        associationResponseParser = new Dot11AssociationResponseFrameParser(metrics, anonymizer);
-        probeRequestParser = new Dot11ProbeRequestFrameParser(metrics, anonymizer);
-        probeResponseFrameParser = new Dot11ProbeResponseFrameParser(metrics, anonymizer);
-        disassociationParser = new Dot11DisassociationFrameParser(metrics, anonymizer);
-        authenticationFrameParser = new Dot11AuthenticationFrameParser(metrics, anonymizer);
-        deauthenticationFrameParser = new Dot11DeauthenticationFrameParser(metrics, anonymizer);
+        this.frameFactory = new Dot11FrameFactory(metrics, anonymizer);
+        this.frameProcessor = frameProcessor;
 
         // Metrics.
         this.globalFrameMeter = metrics.meter(MetricNames.FRAME_COUNT);
@@ -209,7 +182,7 @@ public class Dot11MonitorProbe extends Dot11Probe {
 
                             if (meta.isMalformed()) {
                                 LOG.trace("Bad checksum. Skipping malformed packet on probe [{}].", getName());
-                                statistics.tickMalformedCountAndNotify(meta);
+                                notifyOfMalformedFrame(meta);
                                 continue;
                             }
 
@@ -219,49 +192,12 @@ public class Dot11MonitorProbe extends Dot11Probe {
 
                             mostRecentFrameTimestamp = DateTime.now();
 
-                            // Intercept and handle frame. TODO this looks fucked from a logic perspective
-                            for (Dot11FrameInterceptor interceptor : frameInterceptors) {
-                                if (interceptor.forSubtype() == type.value()) {
-                                    try {
-                                        switch (type.value()) {
-                                            case Dot11FrameSubtype.ASSOCIATION_REQUEST:
-                                                interceptor.intercept(associationRequestParser.parse(payload, r.getHeader().getRawData(), meta));
-                                                break;
-                                            case Dot11FrameSubtype.ASSOCIATION_RESPONSE:
-                                                interceptor.intercept(associationResponseParser.parse(payload, r.getHeader().getRawData(), meta));
-                                                break;
-                                            case Dot11FrameSubtype.PROBE_REQUEST:
-                                                interceptor.intercept(probeRequestParser.parse(payload, r.getHeader().getRawData(), meta));
-                                                break;
-                                            case Dot11FrameSubtype.PROBE_RESPONSE:
-                                                interceptor.intercept(probeResponseFrameParser.parse(payload, r.getHeader().getRawData(), meta));
-                                                break;
-                                            case Dot11FrameSubtype.BEACON:
-                                                interceptor.intercept(beaconParser.parse(payload, r.getHeader().getRawData(), meta));
-                                                break;
-                                            case Dot11FrameSubtype.DISASSOCIATION:
-                                                interceptor.intercept(disassociationParser.parse(payload, r.getHeader().getRawData(), meta));
-                                                break;
-                                            case Dot11FrameSubtype.AUTHENTICATION:
-                                                interceptor.intercept(authenticationFrameParser.parse(payload, r.getHeader().getRawData(), meta));
-                                                break;
-                                            case Dot11FrameSubtype.DEAUTHENTICATION:
-                                                interceptor.intercept(deauthenticationFrameParser.parse(payload, r.getHeader().getRawData(), meta));
-                                                break;
-                                            default:
-                                                LOG.error("Not parsing unsupported management frame subtype [{}].", type.value());
-                                        }
-                                    } catch (MalformedFrameException e) {
-                                        LOG.info("Skipping malformed frame on probe [{}].", getName(), e);
-                                        getStatistics().tickMalformedCountAndNotify(meta);
-                                    }
-                                }
-                            }
+                            // Intercept and handle frame.
+                            frameProcessor.processDot11Frame(frameFactory.build(type, payload, r.getHeader().getRawData(), meta));
 
-                            statistics.tickFrameCount(meta);
                             time.stop();
                         }
-                    } catch(IllegalArgumentException | ArrayIndexOutOfBoundsException | IllegalRawDataException e) {
+                    } catch(IllegalArgumentException | ArrayIndexOutOfBoundsException | MalformedFrameException e) {
                         LOG.debug("Illegal data received on probe [{}].", getName(), e);
                     } catch(Exception e) {
                         LOG.error("Could not process packet on probe [{}].", getName(), e);
@@ -299,16 +235,6 @@ public class Dot11MonitorProbe extends Dot11Probe {
     }
 
     @Override
-    public void addFrameInterceptor(Dot11FrameInterceptor interceptor) {
-        frameInterceptors.add(interceptor);
-    }
-
-    @Override
-    public List<Dot11FrameInterceptor> getInterceptors() {
-        return frameInterceptors;
-    }
-
-    @Override
     @Nullable
     public DateTime getMostRecentFrameTimestamp() {
         return mostRecentFrameTimestamp;
@@ -318,158 +244,15 @@ public class Dot11MonitorProbe extends Dot11Probe {
         channelHopper.onChannelSwitch(handler);
     }
 
-    public static void configureAsBroadMonitor(final Dot11MonitorProbe probe, RemoteConnector uplink) {
-        probe.addFrameInterceptor(new Dot11FrameInterceptor<Dot11AssociationRequestFrame>() {
-            private final Dot11FrameHandler<Dot11AssociationRequestFrame> handler = new Dot11AssociationRequestFrameHandler(probe, uplink);
+    private void notifyOfMalformedFrame(Dot11MetaInformation meta) {
+        int channel = 0;
+        if(meta != null) {
+            channel = meta.getChannel();
+        }
 
-            @Override
-            public void intercept(Dot11AssociationRequestFrame frame) {
-                handler.handle(frame);
-            }
-
-            @Override
-            public byte forSubtype() {
-                return Dot11FrameSubtype.ASSOCIATION_REQUEST;
-            }
-
-            @Override
-            public List<Class<? extends Alert>> raisesAlerts() {
-                return Collections.emptyList();
-            }
-        });
-
-        probe.addFrameInterceptor(new Dot11FrameInterceptor<Dot11AssociationResponseFrame>() {
-            private final Dot11FrameHandler<Dot11AssociationResponseFrame> handler = new Dot11AssociationResponseFrameHandler(probe, uplink);
-
-            @Override
-            public void intercept(Dot11AssociationResponseFrame frame) {
-                handler.handle(frame);
-            }
-
-            @Override
-            public byte forSubtype() {
-                return Dot11FrameSubtype.ASSOCIATION_RESPONSE;
-            }
-
-            @Override
-            public List<Class<? extends Alert>> raisesAlerts() {
-                return Collections.emptyList();
-            }
-        });
-
-        probe.addFrameInterceptor(new Dot11FrameInterceptor<Dot11ProbeRequestFrame>() {
-            private final Dot11FrameHandler<Dot11ProbeRequestFrame> handler = new Dot11ProbeRequestFrameHandler(probe, uplink);
-
-            @Override
-            public void intercept(Dot11ProbeRequestFrame frame) {
-                handler.handle(frame);
-            }
-
-            @Override
-            public byte forSubtype() {
-                return Dot11FrameSubtype.PROBE_REQUEST;
-            }
-
-            @Override
-            public List<Class<? extends Alert>> raisesAlerts() {
-                return Collections.emptyList();
-            }
-        });
-
-        probe.addFrameInterceptor(new Dot11FrameInterceptor<Dot11ProbeResponseFrame>() {
-            private final Dot11FrameHandler<Dot11ProbeResponseFrame> handler = new Dot11ProbeResponseFrameHandler(probe, uplink);
-
-            @Override
-            public void intercept(Dot11ProbeResponseFrame frame) {
-                handler.handle(frame);
-            }
-
-            @Override
-            public byte forSubtype() {
-                return Dot11FrameSubtype.PROBE_RESPONSE;
-            }
-
-            @Override
-            public List<Class<? extends Alert>> raisesAlerts() {
-                return Collections.emptyList();
-            }
-        });
-
-        probe.addFrameInterceptor(new Dot11FrameInterceptor<Dot11BeaconFrame>() {
-            private final Dot11FrameHandler<Dot11BeaconFrame> handler = new Dot11BeaconFrameHandler(probe, uplink);
-
-            @Override
-            public void intercept(Dot11BeaconFrame frame) {
-                handler.handle(frame);
-            }
-
-            @Override
-            public byte forSubtype() {
-                return Dot11FrameSubtype.BEACON;
-            }
-
-            @Override
-            public List<Class<? extends Alert>> raisesAlerts() {
-                return Collections.emptyList();
-            }
-        });
-
-        probe.addFrameInterceptor(new Dot11FrameInterceptor<Dot11DisassociationFrame>() {
-            private final Dot11FrameHandler<Dot11DisassociationFrame> handler = new Dot11DisassociationFrameHandler(probe, uplink);
-
-            @Override
-            public void intercept(Dot11DisassociationFrame frame) {
-                handler.handle(frame);
-            }
-
-            @Override
-            public byte forSubtype() {
-                return Dot11FrameSubtype.DISASSOCIATION;
-            }
-
-            @Override
-            public List<Class<? extends Alert>> raisesAlerts() {
-                return Collections.emptyList();
-            }
-        });
-
-        probe.addFrameInterceptor(new Dot11FrameInterceptor<Dot11AuthenticationFrame>() {
-            private final Dot11FrameHandler<Dot11AuthenticationFrame> handler = new Dot11AuthenticationFrameHandler(probe, uplink);
-
-            @Override
-            public void intercept(Dot11AuthenticationFrame frame) {
-                handler.handle(frame);
-            }
-
-            @Override
-            public byte forSubtype() {
-                return Dot11FrameSubtype.AUTHENTICATION;
-            }
-
-            @Override
-            public List<Class<? extends Alert>> raisesAlerts() {
-                return Collections.emptyList();
-            }
-        });
-
-        probe.addFrameInterceptor(new Dot11FrameInterceptor<Dot11DeauthenticationFrame>() {
-            private final Dot11FrameHandler<Dot11DeauthenticationFrame> handler = new Dot11DeauthenticationFrameHandler(probe, uplink);
-
-            @Override
-            public void intercept(Dot11DeauthenticationFrame frame) {
-                handler.handle(frame);
-            }
-
-            @Override
-            public byte forSubtype() {
-                return Dot11FrameSubtype.DEAUTHENTICATION;
-            }
-
-            @Override
-            public List<Class<? extends Alert>> raisesAlerts() {
-                return Collections.emptyList();
-            }
-        });
+        remote.notifyUplinks(
+                new Notification("Malformed frame received.", channel)
+                        .addField(FieldNames.SUBTYPE, "malformed"), meta);
     }
 
 }
