+++ src/main/java/horse/wtf/nzyme/dot11/probes/Dot11SenderProbe.java
@@ -20,17 +20,13 @@ package horse.wtf.nzyme.dot11.probes;
 import com.codahale.metrics.MetricRegistry;
 import com.google.common.base.Joiner;
 import horse.wtf.nzyme.channels.ChannelHopper;
-import horse.wtf.nzyme.dot11.Dot11FrameInterceptor;
 import horse.wtf.nzyme.dot11.deception.traps.Trap;
-import horse.wtf.nzyme.statistics.Statistics;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.joda.time.DateTime;
 import org.pcap4j.core.*;
 
 import javax.annotation.Nullable;
-import java.util.Collections;
-import java.util.List;
 import java.util.concurrent.atomic.AtomicBoolean;
 
 public class Dot11SenderProbe extends Dot11Probe {
@@ -49,8 +45,8 @@ public class Dot11SenderProbe extends Dot11Probe {
 
     private DateTime mostRecentFrameTimestamp;
 
-    public Dot11SenderProbe(Dot11ProbeConfiguration configuration, Trap trap, Statistics statistics, MetricRegistry metrics) {
-        super(configuration, statistics, metrics);
+    public Dot11SenderProbe(Dot11ProbeConfiguration configuration, Trap trap, MetricRegistry metrics) {
+        super(configuration, metrics);
 
         this.trap = trap;
         this.configuration = configuration;
@@ -154,16 +150,6 @@ public class Dot11SenderProbe extends Dot11Probe {
 
     public Trap getTrap() {
         return trap;
-    }
-
-    @Override
-    public void addFrameInterceptor(Dot11FrameInterceptor interceptor) {
-        throw new RuntimeException("Sender probe cannot intercept frames.");
-    }
-
-    @Override
-    public List<Dot11FrameInterceptor> getInterceptors() {
-        return Collections.emptyList();
     }
 
     @Override
