+++ src/main/protobuf/nzyme_message.proto
@@ -15,29 +15,41 @@
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
 
-package horse.wtf.nzyme.dot11.handlers;
+syntax = "proto3";
 
-import horse.wtf.nzyme.dot11.probes.Dot11Probe;
+option java_outer_classname = "NzymeMessage";
+option java_package = "horse.wtf.nzyme.remote.protobuf";
 
-public abstract class Dot11FrameHandler<T> {
+message Message {
+  string MessageType = 1;
+  Frame frame = 2;
+}
 
-    protected Dot11Probe probe;
+message Frame {
+  string frameType = 1;
+  Dot11Frame dot11Frame = 2;
+}
 
-    protected Dot11FrameHandler(Dot11Probe probe) {
-        this.probe = probe;
-    }
+message Dot11Frame {
 
-    private void tick() {
-        probe.getStatistics().tickType(getName());
-    }
-
-    public void handle(T frame) {
-        tick();
-
-        doHandle(frame);
-    }
-
-    protected abstract void doHandle(T frame);
-    public abstract String getName();
+  int64 recordedAt = 1;
+  string source = 2;
+  string frameType = 3;
+  bytes frameHeader = 4;
+  bytes framePayload = 5;
+  FrameMeta frameMeta = 6;
 
 }
+
+message FrameMeta {
+
+  bool isMalformed = 1;
+  int32 antennaSignal = 2;
+  int32 signalQuality = 3;
+  int32 frequency = 4;
+  int32 channel = 5;
+  int64 macTimestamp = 6;
+  bool isWEP = 7;
+
+}
+
